refactor: split Server.cs (1,594→1,168 lines) — extract order processing to partial class
Extracted all order dispatch, receive, interpret, and sync logic
from Server.cs into Server.OrderProcessing.cs using a C# partial
class. This keeps the same sealed class structure with zero
behavioral change — all methods remain members of the Server class.

Server.cs (1,168 lines): core server lifecycle, connection handling,
   lobby sync, client management, game start/shutdown
Server.OrderProcessing.cs (450 lines): frame packet builders,
   order dispatch/receive, sync hash handling, desync detection,
   server order interpretation (command switch), game save loading

The split follows natural seams already hinted at by TODO comments
about separating order buffering, dynamic latency, and sync logic.
